Re-organized adapters and services. Implemented logout adapter
