+++ src/App/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter as Router } from 'react-router-dom';
 import Navbar from '@components/Navbar';
-import { useHandleCheckAuth } from './AppService';
+import { useHandleCheckAuth } from '@services/authService';
 import Switch from './Switch';
 import Modals from './Modals';
 import './App.css';
