+++ src/services/adapters/types.ts
